tag V00-09-01: fix for slc5_amd64_gcc434 arch.

+++ plugins/MatchRecToGen.cc
@@ -119,8 +119,8 @@ void MatchRecToGen::produce(edm::Event& iEvent,const edm::EventSetup& iSetup)
   iEvent.getByLabel(srcRec_,rec_);
   iEvent.getByLabel(srcGen_,gen_);
   
-  nRec = std::min(rec_->size(),(size_t)100);
-  nGen = std::min(gen_->size(),(size_t)100);
+  nRec = std::min((size_t)rec_->size(),(size_t)100);
+  nGen = std::min((size_t)gen_->size(),(size_t)100);
   
   IndexSet_t iRecSet;
   IndexSet_t iGenSet;
